[FIX] 권장 답변 삭제 시, post 서버 / 캐시 데이터 업데이트
