+++ client/src/components/community/comment/Comment.jsx
@@ -99,7 +99,7 @@ const Comment = ({
               variant="transparent"
               iconSize={20}
               c="var(--font-color)"
-              onClick={() => removeMutate({ commentId: id })}
+              onClick={() => removeMutate({ commentId: id, certified })}
             />
           )}
         </CommentHeader>
